requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
discord.py==1.7.3
future==0.18.2
idna==3.2
multidict==5.1.0
pefile==2021.5.24
Pillow==8.3.1
pip==21.1.3
pycparser==2.20
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
PyNaCl==1.4.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PySide2==5.15.2
pywin32-ctypes==0.2.0
requests==2.26.0
setuptools==56.0.0
shiboken2==5.15.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wheel==0.36.2
yarl==1.6.3
youtube-dl==2021.6.6